+++ src/Treatment.cpp
@@ -187,7 +187,7 @@ void TInfiltro::InitResult(unsigned long nTime, int nInterval, int nCount)
 	m_pResult->m_Header.nData = FT_USER;
 	m_pResult->m_Header.nCount = nCount;
 
-	for(int nCol = 0; nCol < sizeof(pszItemHeader) / sizeof(char*); nCol++)
+	for(int nCol = 0; nCol < (int)(sizeof(pszItemHeader) / sizeof(char*)); nCol++)
 	{
 		TSeriesItem *pItem = m_pResult->AddItem();
 		strcpy_s(pItem->m_Header.szHeader, 50, pszItemHeader[nCol]);
@@ -313,7 +313,7 @@ void TBioRetention::InitResult(unsigned long nTime, int nInterval, int nCount)
 	m_pResult->m_Header.nData = FT_USER;
 	m_pResult->m_Header.nCount = nCount;
 
-	for(int nCol = 0; nCol < sizeof(pszItemHeader) / sizeof(char*); nCol++)
+	for(int nCol = 0; nCol < (int)(sizeof(pszItemHeader) / sizeof(char*)); nCol++)
 	{
 		TSeriesItem *pItem = m_pResult->AddItem();
 		strcpy_s(pItem->m_Header.szHeader, 50, pszItemHeader[nCol]);
@@ -478,7 +478,7 @@ void TWetLand::InitResult(unsigned long nTime, int nInterval, int nCount)
 	m_pResult->m_Header.nData = FT_USER;
 	m_pResult->m_Header.nCount = nCount;
 
-	for(int nCol = 0; nCol < sizeof(pszItemHeader) / sizeof(char*); nCol++)
+	for(int nCol = 0; nCol < (int)(sizeof(pszItemHeader) / sizeof(char*)); nCol++)
 	{
 		TSeriesItem *pItem = m_pResult->AddItem();
 		strcpy_s(pItem->m_Header.szHeader, 50, pszItemHeader[nCol]);
@@ -717,7 +717,7 @@ void TPond::InitResult(unsigned long nTime, int nInterval, int nCount)
 	m_pResult->m_Header.nData = FT_USER;
 	m_pResult->m_Header.nCount = nCount;
 
-	for(int nCol = 0; nCol < sizeof(pszItemHeader) / sizeof(char*); nCol++)
+	for(int nCol = 0; nCol < (int)(sizeof(pszItemHeader) / sizeof(char*)); nCol++)
 	{
 		TSeriesItem *pItem = m_pResult->AddItem();
 		strcpy_s(pItem->m_Header.szHeader, 50, pszItemHeader[nCol]);
@@ -1134,7 +1134,7 @@ void TRainTank::InitResult(unsigned long nTime, int nInterval, int nCount)
 	m_pResult->m_Header.nData = FT_USER;
 	m_pResult->m_Header.nCount = nCount;
 
-	for(int nCol = 0; nCol < sizeof(pszItemHeader) / sizeof(char*); nCol++)
+	for(int nCol = 0; nCol < (int)(sizeof(pszItemHeader) / sizeof(char*)); nCol++)
 	{
 		TSeriesItem *pItem = m_pResult->AddItem();
 		strcpy_s(pItem->m_Header.szHeader, 50, pszItemHeader[nCol]);
@@ -1260,7 +1260,7 @@ void TRecycle::InitResult(unsigned long nTime, int nInterval, int nCount)
 	m_pResult->m_Header.nData = FT_USER;
 	m_pResult->m_Header.nCount = nCount;
 
-	for(nCol = 0; nCol < sizeof(pszItemHeader) / sizeof(char*); nCol++)
+	for(nCol = 0; nCol < (int)(sizeof(pszItemHeader) / sizeof(char*)); nCol++)
 	{
 		TSeriesItem *pItem = m_pResult->AddItem();
 		strcpy_s(pItem->m_Header.szHeader, 50, pszItemHeader[nCol]);
